requirements.txt
kafka-python-ng
psycopg2-binary
python-dotenv

